refactor: extract bundled + dynamic loads from content-fetcher
Phase 3 of the content-fetcher refactor (Pattern J — contract-surface
extraction). Move fetchBundledInteractive and SAFE_PACKAGE_PATH out of
src/docs-retrieval/content-fetcher.ts into a new file
src/docs-retrieval/bundled-loader.ts.

The new module owns:
- the `bundled:` URL scheme and its 4 sub-prefixes
  (wysiwyg-preview / e2e-test / pr-tests/<id> / <dir>/<file>.json)
- the legacy `bundled:<id>` path that consults
  bundled-interactives/index.json for filename resolution
- the SAFE_PACKAGE_PATH whitelist regex (hoisted to module scope so
  it isn't re-allocated per call; behavior preserved verbatim)

Critical contracts preserved verbatim (Pattern J — no rename, no
clean-up; "ownership-path update" only):
- 3 storage-key reads (StorageKeys.WYSIWYG_PREVIEW_JSON,
  StorageKeys.E2E_TEST_GUIDE, sessionStorage['pathfinder-bundled-<id>']).
  StorageKeys constants stay in src/lib/user-storage.ts.
- All 4 error message strings (E2E tests may match them).
- index.json filename fallback: `interactive.filename || `${contentId}.json``.
- Fallback order between the package-path branch and the index.json branch.

Webpack require() literal-path constraint preserved:
src/docs-retrieval/bundled-loader.ts → ../bundled-interactives/...
resolves to src/bundled-interactives/... — same depth as content-fetcher.ts.

content-fetcher.ts changes:
- removes the entire fetchBundledInteractive function (lines 394-632
  in the pre-extraction file)
- drops the now-unused StorageKeys import
- adds `import { fetchBundledInteractive } from './bundled-loader'`

The function is invoked exactly once, from fetchContent's
`bundled:`-prefix branch (same call site as before).

No public API changes: fetchBundledInteractive was never on the
docs-retrieval barrel and has zero external consumers (Grep
confirmed). No consumer migrations needed.

All 13 pre-extraction characterization assertions still pass against
the new module. Full docs-retrieval test slice passes (156 tests
across 10 suites). typecheck clean.

Refs: .refactor/PLAN.md Phase 3, INVESTIGATION §6 invariant 7
(webpack require literal-path), DR-04 (no new barrel exports).
