+++ src/docs-retrieval/content-fetcher.ts
@@ -24,12 +24,12 @@ import {
   isGitHubRawUrl,
 } from '../security';
 import { isDevModeEnabledGlobal } from '../utils/dev-mode';
-import { StorageKeys } from '../lib/user-storage';
 import { generateJourneyContentWithExtras } from './learning-journey-helpers';
 import { resolveRelativeUrls } from './resolve-relative-urls';
 import { validateGuide } from '../validation';
 import { simpleMarkdownToHtml, wrapExpectBlockInOrangeOutline } from './markdown-renderer';
 import { extractMetadata } from './metadata-extractor';
+import { fetchBundledInteractive } from './bundled-loader';
 
 // Internal error structure for detailed error handling
 interface FetchError {
@@ -385,249 +385,6 @@ function generateUserFriendlyError(error: FetchError | undefined, url: string): 
       return 'Server error occurred. Please try again later.';
     default:
       return error.message || 'Failed to load content. Please try again.';
-  }
-}
-
-/**
- * Fetch bundled interactive content from local files
- */
-async function fetchBundledInteractive(url: string): Promise<ContentFetchResult> {
-  const contentId = url.replace('bundled:', '');
-
-  // SPECIAL CASE: Handle WYSIWYG preview from localStorage
-  if (contentId === 'wysiwyg-preview') {
-    try {
-      // Load JSON content from the dedicated preview key
-      const previewContent = localStorage.getItem(StorageKeys.WYSIWYG_PREVIEW_JSON);
-
-      if (!previewContent || previewContent.trim() === '') {
-        return {
-          content: null,
-          error: 'No preview content available. Create content in the WYSIWYG editor and click Test first.',
-        };
-      }
-
-      // Content is already JSON - parse to extract title for metadata
-      let title = 'Preview: WYSIWYG Guide';
-      try {
-        const parsed = JSON.parse(previewContent);
-        if (parsed.title) {
-          title = parsed.title;
-        }
-      } catch {
-        // If parsing fails, use default title
-      }
-
-      const rawContent: RawContent = {
-        content: previewContent, // Already valid JSON guide format
-        metadata: {
-          title,
-        },
-        type: 'interactive',
-        url,
-        lastFetched: new Date().toISOString(),
-      };
-
-      return { content: rawContent };
-    } catch (error) {
-      console.error('Failed to load WYSIWYG preview:', error);
-      return {
-        content: null,
-        error: `Failed to load preview: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      };
-    }
-  }
-
-  // SPECIAL CASE: Handle E2E test guide from localStorage
-  // This is used by the E2E test runner CLI to inject arbitrary JSON guides for testing
-  if (contentId === 'e2e-test') {
-    try {
-      // Load JSON content from the E2E test key
-      const testContent = localStorage.getItem(StorageKeys.E2E_TEST_GUIDE);
-
-      if (!testContent || testContent.trim() === '') {
-        return {
-          content: null,
-          error: 'No E2E test content available. The E2E runner must inject JSON into localStorage first.',
-        };
-      }
-
-      // Content is already JSON - parse to extract title for metadata
-      let title = 'E2E Test Guide';
-      try {
-        const parsed = JSON.parse(testContent);
-        if (parsed.title) {
-          title = parsed.title;
-        }
-      } catch {
-        // If parsing fails, use default title
-      }
-
-      const rawContent: RawContent = {
-        content: testContent, // Already valid JSON guide format
-        metadata: {
-          title,
-        },
-        type: 'interactive',
-        url,
-        lastFetched: new Date().toISOString(),
-      };
-
-      return { content: rawContent };
-    } catch (error) {
-      console.error('Failed to load E2E test guide:', error);
-      return {
-        content: null,
-        error: `Failed to load E2E test guide: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      };
-    }
-  }
-
-  // SPECIAL CASE: Handle PR test learning paths from sessionStorage
-  if (contentId.startsWith('pr-tests/')) {
-    try {
-      const pathId = contentId.replace('pr-tests/', '');
-      const storageKey = `pathfinder-bundled-${pathId}`;
-      const pathContent = sessionStorage.getItem(storageKey);
-
-      if (!pathContent || pathContent.trim() === '') {
-        return {
-          content: null,
-          error: 'PR test path not found. It may have expired or been cleared.',
-        };
-      }
-
-      // Content is already JSON - parse to extract title for metadata
-      let title = 'PR Test Path';
-      try {
-        const parsed = JSON.parse(pathContent);
-        if (parsed.title) {
-          title = parsed.title;
-        }
-      } catch {
-        // If parsing fails, use default title
-      }
-
-      const rawContent: RawContent = {
-        content: pathContent, // Already valid JSON guide format
-        metadata: {
-          title,
-        },
-        type: 'interactive',
-        url,
-        lastFetched: new Date().toISOString(),
-      };
-
-      return { content: rawContent };
-    } catch (error) {
-      console.error('Failed to load PR test path:', error);
-      return {
-        content: null,
-        error: `Failed to load test path: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      };
-    }
-  }
-
-  // Handle package-format paths: bundled:<package-dir>/content.json
-  // These are produced by BundledPackageResolver and follow the two-file package model.
-  const SAFE_PACKAGE_PATH = /^[a-z0-9][a-z0-9-]*\/[a-z0-9][a-z0-9._-]*\.json$/;
-  if (contentId.includes('/') && contentId.endsWith('.json')) {
-    if (!SAFE_PACKAGE_PATH.test(contentId)) {
-      return {
-        content: null,
-        error: `Invalid bundled package path: ${contentId}`,
-        errorType: 'not-found',
-      };
-    }
-
-    try {
-      const jsonModule = require(`../bundled-interactives/${contentId}`);
-      const jsonContent = typeof jsonModule === 'string' ? jsonModule : JSON.stringify(jsonModule);
-
-      if (!jsonContent || jsonContent.trim() === '' || jsonContent === '{}') {
-        return {
-          content: null,
-          error: `Bundled package file not found: ${contentId}`,
-          errorType: 'not-found',
-        };
-      }
-
-      // Webpack imports JSON as objects — read title directly to avoid a round-trip
-      const moduleTitle =
-        typeof jsonModule === 'object' && jsonModule !== null && typeof jsonModule.title === 'string'
-          ? jsonModule.title
-          : undefined;
-      const title: string = moduleTitle ?? contentId.split('/')[0] ?? contentId;
-
-      const rawContent: RawContent = {
-        content: jsonContent,
-        metadata: { title },
-        type: 'interactive',
-        url,
-        lastFetched: new Date().toISOString(),
-        isNativeJson: true,
-      };
-
-      return { content: rawContent };
-    } catch (err) {
-      console.warn(`[docs-retrieval] Failed to load bundled package file: ${contentId}`, err);
-      return {
-        content: null,
-        error: `Bundled package file not found: ${contentId}`,
-        errorType: 'not-found',
-      };
-    }
-  }
-
-  // Load bundled interactive from index.json
-  // JSON format is the standard - all bundled interactives should be .json files
-  try {
-    // Load the index.json to find the correct filename for this interactive
-    const indexData = require('../bundled-interactives/index.json');
-    const interactive = indexData?.interactives?.find((item: any) => item.id === contentId);
-
-    if (!interactive) {
-      return {
-        content: null,
-        error: `Bundled interactive not found in index.json: ${contentId}`,
-      };
-    }
-
-    // Load JSON guide (standard format)
-    const filename = interactive.filename || `${contentId}.json`;
-    const jsonModule = require(`../bundled-interactives/${filename}`);
-
-    // JSON files are imported as objects by webpack, stringify for consistent handling
-    const jsonContent = typeof jsonModule === 'string' ? jsonModule : JSON.stringify(jsonModule);
-
-    if (!jsonContent || jsonContent.trim() === '' || jsonContent === '{}') {
-      return {
-        content: null,
-        error: `Bundled interactive content is empty: ${contentId}`,
-      };
-    }
-
-    // For JSON guides, we store the JSON string in the content field
-    // The ContentProcessor will detect and parse it appropriately
-    const rawContent: RawContent = {
-      content: jsonContent,
-      metadata: {
-        title: interactive.title || contentId,
-      },
-      type: 'interactive',
-      url,
-      lastFetched: new Date().toISOString(),
-    };
-
-    return { content: rawContent };
-  } catch (error) {
-    console.error(`Failed to load bundled interactive ${contentId}:`, error);
-    return {
-      content: null,
-      error: `Failed to load bundled interactive: ${contentId}. Error: ${
-        error instanceof Error ? error.message : 'Unknown error'
-      }`,
-    };
   }
 }
